+++ src/par_iter/reduce.rs
@@ -101,7 +101,7 @@ impl<PAR_ITER, REDUCE_OP> Consumer for ReduceConsumer<PAR_ITER, REDUCE_OP>
     type SeqState = PAR_ITER::Item;
     type Result = PAR_ITER::Item;
 
-    unsafe fn split_at(self, _index: u64) -> (Self, Self) {
+    unsafe fn split_at(self, _index: usize) -> (Self, Self) {
         (ReduceConsumer::new(), ReduceConsumer::new())
     }
